# Reunify the custom and built-in handlers collections to avoid computing the collections per-request