--- a/src/OpenIddict.Server.AspNetCore/OpenIddictServerAspNetCoreHandler.cs
+++ b/src/OpenIddict.Server.AspNetCore/OpenIddictServerAspNetCoreHandler.cs
@@ -29,19 +29,24 @@ namespace OpenIddict.Server.AspNetCore
         IAuthenticationSignInHandler,
         IAuthenticationSignOutHandler
     {
-        private readonly IOpenIddictServerProvider _provider;
+        private readonly IOpenIddictServerDispatcher _dispatcher;
+        private readonly IOpenIddictServerFactory _factory;
 
         /// <summary>
         /// Creates a new instance of the <see cref="OpenIddictServerAspNetCoreHandler"/> class.
         /// </summary>
         public OpenIddictServerAspNetCoreHandler(
-            [NotNull] IOpenIddictServerProvider provider,
+            [NotNull] IOpenIddictServerDispatcher dispatcher,
+            [NotNull] IOpenIddictServerFactory factory,
             [NotNull] IOptionsMonitor<OpenIddictServerAspNetCoreOptions> options,
             [NotNull] ILoggerFactory logger,
             [NotNull] UrlEncoder encoder,
             [NotNull] ISystemClock clock)
             : base(options, logger, encoder, clock)
-            => _provider = provider;
+        {
+            _dispatcher = dispatcher;
+            _factory = factory;
+        }
 
         public async Task<bool> HandleRequestAsync()
         {
@@ -51,7 +56,7 @@ namespace OpenIddict.Server.AspNetCore
             if (transaction == null)
             {
                 // Create a new transaction and attach the HTTP request to make it available to the ASP.NET Core handlers.
-                transaction = await _provider.CreateTransactionAsync();
+                transaction = await _factory.CreateTransactionAsync();
                 transaction.Properties[typeof(HttpRequest).FullName] = new WeakReference<HttpRequest>(Request);
 
                 // Attach the OpenIddict server transaction to the ASP.NET Core features
@@ -60,7 +65,7 @@ namespace OpenIddict.Server.AspNetCore
             }
 
             var context = new ProcessRequestContext(transaction);
-            await _provider.DispatchAsync(context);
+            await _dispatcher.DispatchAsync(context);
 
             if (context.IsRequestHandled)
             {
@@ -84,7 +89,7 @@ namespace OpenIddict.Server.AspNetCore
                     }
                 };
 
-                await _provider.DispatchAsync(notification);
+                await _dispatcher.DispatchAsync(notification);
 
                 if (notification.IsRequestHandled)
                 {
@@ -118,7 +123,7 @@ namespace OpenIddict.Server.AspNetCore
             if (context == null)
             {
                 context = new ProcessAuthenticationContext(transaction);
-                await _provider.DispatchAsync(context);
+                await _dispatcher.DispatchAsync(context);
 
                 // Store the context object in the transaction so it can be later retrieved by handlers
                 // that want to access the authentication result without triggering a new authentication flow.
@@ -182,7 +187,7 @@ namespace OpenIddict.Server.AspNetCore
                 Response = new OpenIddictResponse()
             };
 
-            await _provider.DispatchAsync(context);
+            await _dispatcher.DispatchAsync(context);
 
             if (context.IsRequestHandled || context.IsRequestSkipped)
             {
@@ -201,7 +206,7 @@ namespace OpenIddict.Server.AspNetCore
                     }
                 };
 
-                await _provider.DispatchAsync(notification);
+                await _dispatcher.DispatchAsync(notification);
 
                 if (notification.IsRequestHandled || context.IsRequestSkipped)
                 {
@@ -237,7 +242,7 @@ namespace OpenIddict.Server.AspNetCore
                 Response = new OpenIddictResponse()
             };
 
-            await _provider.DispatchAsync(context);
+            await _dispatcher.DispatchAsync(context);
 
             if (context.IsRequestHandled || context.IsRequestSkipped)
             {
@@ -256,7 +261,7 @@ namespace OpenIddict.Server.AspNetCore
                     }
                 };
 
-                await _provider.DispatchAsync(notification);
+                await _dispatcher.DispatchAsync(notification);
 
                 if (notification.IsRequestHandled || context.IsRequestSkipped)
                 {
@@ -283,7 +288,7 @@ namespace OpenIddict.Server.AspNetCore
 
             transaction.Properties[typeof(AuthenticationProperties).FullName] = properties ?? new AuthenticationProperties();
 
-            await _provider.DispatchAsync(context);
+            await _dispatcher.DispatchAsync(context);
 
             if (context.IsRequestHandled || context.IsRequestSkipped)
             {
@@ -302,7 +307,7 @@ namespace OpenIddict.Server.AspNetCore
                     }
                 };
 
-                await _provider.DispatchAsync(notification);
+                await _dispatcher.DispatchAsync(notification);
 
                 if (notification.IsRequestHandled || context.IsRequestSkipped)
                 {