In beta1, the OpenIddict server and validation components build the handlers list per-request by combining the built-in handlers registered via `options.Handlers` and the user-defined handlers added to `options.CustomHandlers` and by sorting the resulting list using the order assigned to each descriptor. This logic is robust but extremely expensive. This PR moves away from this pattern and merges the two handlers collection into a single one, sorted at startup rather than per-request.

The perf' boost is impressive: **almost +33% reqs/sec**.

Before:

```
Bombarding http://localhost:54540/.well-known/openid-configuration for 1m0s using 125 connection(s)
Done!
Statistics        Avg      Stdev        Max
  Reqs/sec     50132.71    7227.46   74182.20
  Latency        2.48ms   132.92us    41.00ms
  HTTP codes:
    1xx - 0, 2xx - 3023763, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:    63.97MB/s
```

After:

```
Bombarding http://localhost:54540/.well-known/openid-configuration for 1m0s using 125 connection(s)
Done!
Statistics        Avg      Stdev        Max
  Reqs/sec     66466.17    8644.19   83914.57
  Latency        1.87ms   109.27us    45.00ms
  HTTP codes:
    1xx - 0, 2xx - 3999942, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:    84.62MB/s
```

Note: this PR also renames `IOpenIddict(Server/Validation)Provider` to `IOpenIddict(Server/Validation)Dispatcher` and moves the `CreateTransactionAsync()` method to a separate factory.